Library

SparseIdentification.CompoundBasis — Type

CompoundBasis(bases)

Holds a basis composed of several different basis functions, e.g. polynomials of vardatag degree and/or trigonometric functions.

SparseIdentification.HamiltonianSINDyVectorField — Method

wrapper function for generalized SINDY hamiltonian gradient. Needs the output of fθ_sparse to work! It is in a syntax that is suitable to be evaluated by a loss function for optimization

SparseIdentification.PolynomialBasis — Type

PolynomialBasis(p)

Holds polynomials of degree p.

SparseIdentification.TrigonometricBasis — Type

TrigonometricBasis(n)

Holds the basis functions [sin(kx), cos(kx)] for 1 ≤ k ≤ n.

SparseIdentification.calculate_nparams — Method

returns the number of required parameters depending on whether there are trig basis or not

SparseIdentification.evaluate — Function

evaluate(data, basis)

Evaluates a data set on all basis functions in basis.

SparseIdentification.hamilGrad_func_builder — Method

returns a function that can build the gradient of the hamiltonian

SparseIdentification.hamil_trig — Method

collects and sums polynomial and trignometric combinations of basis

SparseIdentification.hamiltonian — Method

collects and sums only polynomial combinations of basis

SparseIdentification.hamiltonian_poly — Method

makes polynomial combinations of basis

SparseIdentification.sparsify — Method

sequential least squares